+++ apps/frontend/components/faq/FaqAnnulerReservation.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    <p class="mb-3">
+      Bien sûr. Note cependant que ton nombre de Coins sera quand même augmenté en fonction du temps
+      écoulé entre la validation et l'annulation (Voir paragraphe "Comment sont comptés les Coins
+      ?"). En d'autres termes, plus tu annules tôt moins tu seras pénalisé parce que cela laisse
+      plus de temps à un autre pilote pour s'organiser.
+    </p>
+    <p class="mt-3">
+      Privilégie les demandes sur des dates où tu es vraiment certain d'être disponible pour voler
+      pour éviter ce désagrément.
+    </p>
+  </div>
+</template>
