+++ apps/frontend/components/faq/FaqDemandeNonValidee.vue
@@ -1,0 +1,17 @@
+<template>
+  <div>
+    <p class="mb-3">
+      Si ta demande n'est pas validée à J-6, cela signifie que d'autres pilotes avec un nombre de
+      Coins plus faible ont été prioritaires sur les matériels disponibles.
+    </p>
+    <p class="mb-3"><strong>Tes options :</strong></p>
+    <ul class="list-disc pl-6 space-y-2">
+      <li>
+        Conserver ta demande en attente au cas où des biplaces se libèrent (annulation d'une autre
+        réservation).
+      </li>
+      <li>Annuler ta demande.</li>
+    </ul>
+    <p class="mt-3">En tout état de cause aucun Coin n'est crédité.</p>
+  </div>
+</template>
