+++ apps/frontend/components/faq/FaqDemandeVsReservation.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <p class="mb-3">
+      <strong class="text-gray-900">Une demande de réservation</strong> est ton souhait de disposer
+      d'un biplace pour une date précise. C'est la première étape du processus. Tu exprimes ton
+      intention pour un ou plusieurs packs, mais pour l'instant, aucun ne t'est attribué.
+    </p>
+    <p class="mb-3">
+      <strong class="text-gray-900">Une réservation</strong> est une demande validée par le système.
+      Tu as alors un créneau garanti pour voler avec un pack attribué.
+    </p>
+    <p>
+      Le passage d'une demande à une réservation est réalisée par l'application en assurant une
+      équité maximale entre les pilotes (voir question "comment fonctionne l'algorithme
+      d'attribution des packs ?").
+    </p>
+  </div>
+</template>
