+++ apps/frontend/components/faq/FaqDuckSheet.vue
@@ -1,0 +1,11 @@
+<template>
+  <p class="mb-3">
+    Nous aussi mais cette bonne vieille sheet avait un certain nombre de limites que cette
+    application vise à combler :
+  </p>
+  <ul class="list-disc pl-6 space-y-2">
+    <li>Équité des pilotes pour l'utilisation du matériel.</li>
+    <li>Ergonomie discutable, surtout sur mobile.</li>
+    <li>Carnet de vol parfois non rempli.</li>
+  </ul>
+</template>
