+++ apps/frontend/components/faq/FaqFormationExamenClub.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    <p hidden>qbi préfo pré-formation qualification</p>
+    <p class="mb-3">
+      Ces événements sont prioritaires par rapport au système de Coins. Contacte un administrateur
+      pour qu'il valide ta demande de réservation avant J-6. Le plus tôt possible est le mieux.
+    </p>
+    <p>
+      Après J-6, il te faudra soit postuler sur un matériel disponible soit négocier avec les
+      pilotes plus prioritaires pour en obtenir un.
+    </p>
+  </div>
+</template>
