+++ apps/frontend/components/faq/FaqJoursConsecutifs.vue
@@ -1,0 +1,33 @@
+<template>
+  <div>
+    <p class="mb-3">
+      Aujourd'hui, et probablement pour encore quelques temps, la notion de "réserver plusieurs
+      jours" n'existe pas. Tu fais autant de demandes que de jours désirés et chaque demande est
+      gérée indépendamment des autres.
+    </p>
+    <p class="mb-3">
+      <strong>Exemple :</strong> Tu fais des demandes à J-7 pour samedi et dimanche, il est
+      absolument possible que tu n'aies du matériel que pour samedi, que pour dimanche, pour samedi
+      et dimanche mais il est possible que les matériels soient différents.
+    </p>
+    <p class="mb-3"><strong>Que faire dans ce cas ?</strong></p>
+    <p class="mb-3">
+      Nous sommes un club. Il est toujours temps de discuter pour tenter de trouver un arrangement
+      avec les pilotes plus prioritaires que toi.
+    </p>
+    <p class="mb-3">
+      Par ailleurs, si tu as du matériel affecté pour deux jours mais que ce n'est pas le même,
+      contacte un canard qui a un autre matériel pour échanger l'un des deux jours. En cas d'accord,
+      contacte un admin pour qu'il intervertisse les réservations (nécessaire pour le carnet de
+      vol).
+    </p>
+    <p class="mb-3"><strong>Pro tip :</strong></p>
+    <p>
+      Imaginons que tu désires un biplace samedi et un dimanche. Si le lundi soir après
+      l'attribution de J-6 (à 20:00), tu constates qu'un biplace est disponible pour samedi et
+      dimanche. Dépose une demande pour chaque jour sur ce biplace et il te sera attribué pour les
+      deux jours dans les 30 minutes sous réserve que personne de plus prioritaire ne se soit
+      positionné pendant ces 30 minutes.
+    </p>
+  </div>
+</template>
