+++ apps/frontend/components/faq/FaqMaximiserChances.vue
@@ -1,0 +1,43 @@
+<template>
+  <ul class="list-disc pl-6 space-y-3">
+    <li>
+      <strong>Postule à tous les biplaces qui peuvent te convenir</strong> plutôt qu'à un seulement
+      ton préféré.
+    </li>
+    <li>
+      <strong>Postule aux jours avec une faible demande.</strong> <br />De façon caricaturale : en
+      semaine et ou en hiver. Tu as compris l'idée.<br />
+      Même si tu as beaucoup de Coins, tu peux te placer dans les trous pour voler quand même.
+    </li>
+    <li>
+      <strong>Réserver peu</strong> permet d'avoir plus de choix dans la date et d'augmenter la
+      probabilité d'avoir du matériel plusieurs jours consécutifs (voir plus bas).
+    </li>
+    <li>
+      <strong>En engrangeant le moins de Coins possibles :</strong>
+      <ul class="list-disc pl-6 space-y-3">
+        <li>
+          <strong>Gère tes demandes :</strong><br />
+          Nous sommes le 1er jour du mois, tu as demandé un pack pour le 4 mais aucun pack ne t'a
+          été attribué. <br />Tant qu'un biplace t'intéresse, conserve ta demande pour le 4. En
+          effet, si le pilote à qui il est attribué se désiste, il te sera peut-être attribué.<br />
+          Par contre, si le 2, tu décides que comme tu n'as pas de biplace, tu iras grimper avec tes
+          potes, annule ta demande. Dans le cas contraire, si le pilote à qui il est attribué se
+          désiste, il te sera peut-être attribué et tu seras redevable d'au moins un Coin même si tu
+          l'annules rapidement.
+        </li>
+        <li>
+          <strong>Réserver plus tard que J-6 coûte moins cher mais est plus risqué.</strong><br />
+          Cela peut être une stratégie plus économique dans les périodes de faible demande : Si tu
+          désires voler en février, la demande est généralement faible. En consultant régulièrement
+          le planning, tu peux attendre le dernier moment pour réserver et ainsi limiter le nombre
+          de Coins.
+        </li>
+        <li>
+          <strong>Si tu sens que la météo va tourner au vinaigre,</strong> annule ta réservation le
+          plus tôt possible, cela coûte moins de Coins que la veille du vol ou le jour même.
+        </li>
+      </ul>
+    </li>
+  </ul>
+</template>
