+++ apps/frontend/components/faq/FaqMotDePasse.vue
@@ -1,0 +1,6 @@
+<template>
+  <p class="mb-3">
+    Depuis la page d'accueil, clique sur bouton "Se connecter" puis sur "mot de passe oublié ? ".
+    Renseigne ton email et suis la procédure.
+  </p>
+</template>
