+++ apps/frontend/components/faq/FaqPasPuVole.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <p class="mb-3">
+      Désolé mais les Coins seront quand même comptés. Outre le fait que personne ne peut vérifier
+      tes dires, tu as eu le matériel pour toi alors qu'un autre pilote aurait peut-être fait le
+      choix d'un endroit où il n'a pas plu.<br />
+    </p>
+    <p class="mb-3">
+      L'autre raison est pour limiter les réservations posées "au cas où j'aurais envie" et qui ne
+      sont ni honorées ni annulées. Le principe étant de favoriser l'annulation au plus tôt pour
+      permettre à d'autres pilotes de s'organiser.
+    </p>
+    <p class="mb-3">
+      Sur le plan strict de la météo, nous sommes conscients que cela est frustrant mais tout le
+      monde est soumis aux aléas météo de la même façon, cela est donc équitable dans la durée.
+    </p>
+  </div>
+</template>
