+++ apps/frontend/components/faq/FaqQuestCeQuUnCoin.vue
@@ -1,0 +1,7 @@
+<template>
+  <p>
+    C'est le petit nom donné aux points qui sont attribués aux pilotes pour déterminer leur
+    priorité. <strong>Plus tu as de Coins, moins tu es prioritaire.</strong> Lis la suite pour
+    comprendre comment ils sont calculés.
+  </p>
+</template>
