+++ apps/frontend/components/faq/FaqRecupererMateriel.vue
@@ -1,0 +1,10 @@
+<template>
+  <div>
+    <p hidden>chercher</p>
+    <p class="mb-3">
+      Cette information se trouve sous le menu "Infos packs". Choisis ton pack et tu obtiendras les
+      caractérisiques du pack ainsi que son carnet de vols/réservations dans lequel tu trouveras ce
+      que tu cherches. Bon vol !
+    </p>
+  </div>
+</template>
