+++ apps/frontend/components/faq/FaqScoreCoins.vue
@@ -1,0 +1,6 @@
+<template>
+  <p>
+    Ton score Coins est visible sur ta
+    <NuxtLink to="/mon-compte">page de profil</NuxtLink>. Il est mis à jour en temps réel.
+  </p>
+</template>
